+++ DIII4A/idTech4Amm/src/main/java/com/karin/idTech4Amm/sys/Constants.java
@@ -12,8 +12,8 @@ public final class Constants
     public static final String CONST_PREFERENCE_APP_CRASH_INFO = "_APP_CRASH_INFO";
     public static final String CONST_PREFERENCE_EXCEPTION_DEBUG = "_EXCEPTION_DEBUG";
 
-    public static final int CONST_UPDATE_RELEASE = 28;
-    public static final String CONST_RELEASE = "2023-04-13";
+    public static final int CONST_UPDATE_RELEASE = 29;
+    public static final String CONST_RELEASE = "2023-05-01";
     public static final String CONST_EMAIL = "beyondk2000@gmail.com";
     public static final String CONST_DEV = "Karin";
     public static final String CONST_CODE = "Harmattan";
@@ -25,10 +25,11 @@ public final class Constants
     public static final String CONST_DEVELOPER_XDA = "https://forum.xda-developers.com/member.php?u=10584229";
 	public static final String CONST_CHECK_FOR_UPDATE_URL = "https://raw.githubusercontent.com/glKarin/com.n0n3m4.diii4a/package/CHECK_FOR_UPDATE.json";
 	public static final String[] CONST_CHANGES = {
-            "Add bool cvar `harm_g_mutePlayerFootStep` to control mute player footstep sound(default on) in Quake 4.",
-            "Fix some light's brightness depend on sound amplitude in Quake 4. e.g. in most levels like `airdefense2`, at some dark passages, it should has a repeat-flashing lighting.",
-            "Remove Quake 4 helper dialog when start Quake 4, if want to extract resource files, open `Other` -> `Extract resource` in menu.",
-            "(Bug)In Quake 4, if load some levels has noise with effects on, typed `bse_enable` to 0, and then typed `bse_enable` back to 1 in console, noise can resolved.",
+            "Fixup crash in game loading when change app to background.",
+            "Fixup effects with noise in Quake 4.",
+            "Optimize sky render in Quake 4.",
+            "Remove cvar `harm_g_flashlightOn` in Quake 4.",
+            "Fixup on-screen buttons layer render error on some devices.",
 	};
     public static final String CONST_PACKAGE = "com.karin.idTech4Amm";